+++ database/migrations/2020_04_08_013330_create_awards_categories_table.php
@@ -15,12 +15,12 @@ class CreateAwardsCategoriesTable extends Migration
 	{
 		Schema::connection('former_app_database')->create('awards_categories', function(Blueprint $table)
 		{
-			$table->mediumInteger('id_categorie', true);
-			$table->mediumInteger('id_serie_categorie');
+			$table->integer('id_categorie', true);
+			$table->integer('id_serie_categorie');
 			$table->string('nom_categorie');
 			$table->text('description_categorie');
 			$table->boolean('niveau_categorie')->comment('1: meilleurjeu, 2: general, 3: specialisé, 4: coupdecoeur');
-			$table->mediumInteger('id_session');
+			$table->integer('id_session');
 			$table->dateTime('date_ajout_categorie');
 			$table->smallInteger('statut_categorie');
 			$table->boolean('is_declinaison');
