+++ database/migrations/2020_04_08_013330_create_jukebox_table.php
@@ -15,12 +15,12 @@ class CreateJukeboxTable extends Migration
 	{
 		Schema::connection('former_app_database')->create('jukebox', function(Blueprint $table)
 		{
-			$table->mediumInteger('id', true);
+			$table->integer('id', true);
 			$table->string('titre');
-			$table->mediumInteger('id_posteur');
+			$table->integer('id_posteur');
 			$table->dateTime('date_publication');
 			$table->text('com');
-			$table->mediumInteger('id_jeu_origine');
+			$table->integer('id_jeu_origine');
 			$table->string('url_fichier');
 			$table->boolean('statut_zik')->default(1);
 		});
