+++ database/migrations/2020_04_08_013330_create_news_categories_table.php
@@ -15,7 +15,7 @@ class CreateNewsCategoriesTable extends Migration
 	{
 		Schema::connection('former_app_database')->create('news_categories', function(Blueprint $table)
 		{
-			$table->mediumInteger('id_categorie', true);
+			$table->integer('id_categorie', true);
 			$table->string('titre_categorie');
 			$table->string('titre_categorie_public');
 			$table->string('couleur', 6);
