+++ database/migrations/2020_04_08_013330_create_pre_tests_categories_table.php
@@ -15,10 +15,10 @@ class CreatePreTestsCategoriesTable extends Migration
 	{
 		Schema::connection('former_app_database')->create('pre_tests_categories', function(Blueprint $table)
 		{
-			$table->mediumInteger('id_serie');
-			$table->mediumInteger('id_jeu');
-			$table->mediumInteger('id_jury');
-			$table->mediumInteger('id_categorie');
+			$table->integer('id_serie');
+			$table->integer('id_jeu');
+			$table->integer('id_jury');
+			$table->integer('id_categorie');
 			$table->smallInteger('statut_ptc');
 			$table->primary(['id_serie','id_jeu','id_jury','id_categorie'], 'pre_tests_attributions_primary');
 		});
