+++ database/migrations/2020_04_08_013330_create_recap_table.php
@@ -15,11 +15,11 @@ class CreateRecapTable extends Migration
 	{
 		Schema::connection('former_app_database')->create('recap', function(Blueprint $table)
 		{
-			$table->mediumInteger('id', true);
-			$table->mediumInteger('id_membre');
+			$table->integer('id', true);
+			$table->integer('id_membre');
 			$table->dateTime('date_action');
 			$table->string('ip', 15);
-			$table->mediumInteger('id_element');
+			$table->integer('id_element');
 			$table->smallInteger('type_element')->comment('1:news, 2:jeu, 3:participant, 4:screen, 5:test, 6:serie_test, 7:valid_test, 8:mb, 9:coms, 10:droits, 11:deliberations, 12:pseudo, 13:inscription, 14:acces-jukebox, 15:connexion, 16:nomines');
 			$table->smallInteger('type_action')->comment('1:ajout,2:modif,3:erreur,4:suppr');
 			$table->string('commentaire', 100);
