+++ database/migrations/2020_04_08_013330_create_screenshots_du_moment_table.php
@@ -18,7 +18,7 @@ class CreateScreenshotsDuMomentTable extends Migration
 			$table->smallInteger('id_screenshot', true);
 			$table->string('url_screenshot');
 			$table->dateTime('date');
-			$table->mediumInteger('id_membre');
+			$table->integer('id_membre');
 		});
 	}
